requirements.txt
absl-py==0.8.1
amqp==1.4.9
anyjson==0.3.3
astor==0.8.0
billiard==3.3.0.23
cachetools==3.1.1
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
Django==2.2.5
django-celery==3.2.2
django-suit==0.2.28
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==3.0.37
Markdown==3.1.1
mtcnn==0.1.0
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
PyYAML==5.2
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn
sqlparse==0.3.0
tensorflow-gpu==1.15.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
